+++ frontend/app/src/components/history/events/HistoryEventsListItem.vue
@@ -11,6 +11,8 @@ import HistoryEventNote from '@/components/history/events/HistoryEventNote.vue';
 import HistoryEventsListItemAction from '@/components/history/events/HistoryEventsListItemAction.vue';
 import HistoryEventType from '@/components/history/events/HistoryEventType.vue';
 import { useSupportedChains } from '@/composables/info/chains';
+import { useRefMap } from '@/composables/utils/useRefMap';
+import { useIgnoredAssetsStore } from '@/store/assets/ignored';
 
 const props = defineProps<{
   item: HistoryEventEntry;
@@ -28,6 +30,8 @@ const emit = defineEmits<{
   'show:missing-rule-action': [data: HistoryEventEditData];
   'refresh': [];
 }>();
+
+const { item } = toRefs(props);
 
 const { getChain } = useSupportedChains();
 
@@ -77,6 +81,11 @@ function getEventNoteAttrs(event: HistoryEventEntry) {
     ...data,
   };
 }
+
+const eventAsset = useRefMap(item, ({ asset }) => asset);
+
+const { useIsAssetIgnored } = useIgnoredAssetsStore();
+const isIgnoredAsset = useIsAssetIgnored(eventAsset);
 </script>
 
 <template>
@@ -86,6 +95,7 @@ function getEventNoteAttrs(event: HistoryEventEntry) {
     :class="{
       'bg-rui-error/[0.05]': isHighlighted,
       'border-b': !isLast && !compact,
+      '!opacity-50': isIgnoredAsset,
     }"
   >
     <div